added vet stores to care page, and refactor
