+++ lib/views/care_page.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'dart:core';
+import 'veterinarystores_page.dart';
 
 class Care extends StatelessWidget {
   const Care({Key? key}) : super(key: key);
@@ -8,7 +9,7 @@ class Care extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: _appBar(),
-      body: careView(),
+      body: careView(context),
     );
   }
 
@@ -39,7 +40,7 @@ class Care extends StatelessWidget {
     );
   }
 
-  Widget careView() {
+  Widget careView(BuildContext context) {
     return ListView(
       padding: const EdgeInsets.all(8),
       children: [
@@ -98,6 +99,33 @@ class Care extends StatelessWidget {
                 subtitle: Text('This is the section for pet adoption'),
               ),
             ],
+          ),
+        ),
+        GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(builder: (context) => Search()),
+            );
+          },
+          child: Card(
+            color: Colors.yellow, // set color here
+            child: Column(
+              children: [
+                SizedBox(
+                  height: 150, // set fixed height for image container
+                  width: double.infinity, // set width to fill available space
+                  child: Image.network(
+                    'https://picsum.photos/id/1019/600/300',
+                    fit: BoxFit.cover,
+                  ),
+                ),
+                const ListTile(
+                  title: Text('Veterinary Stores'),
+                  subtitle: Text('This is the section for Veterinary Stores'),
+                ),
+              ],
+            ),
           ),
         ),
       ],
